Add independent files for full BDD support
How to

1. Run full BDD algorithm:

• Run create_parity_automata.sh with the path to the .tlsf input file as first argument, and the name of the .tlsf file as second argument (generate_every_automaton.sh performs this step for an entire folder containing .tlsf files).
• Run spore_full_bdd.py with, as first argument, the newly generated data.txt from the previous command.

Basic arguments are available for spore_full_bdd.py : the second is a boolean that says if the dynamic reordering should be set and the third says if we reorder manually the BDD before the product. Both values are set to 'True' by default.

Inside the file spore_full_bdd.py, the variable 'info' can be set to false so the output will only be "10" or "20" (or error).

2. Run benchmarks:

The file comparison.py could not stop the execution of certain algorithms for timeout, so the file benchmarks.sh replaces it.

• Modify generate_every_automaton.sh and benchmarks.sh, a variable indicates where are the .tlsf files of the benchmarks.
• Run generate_every_automaton.sh.
• Run benchmarks.sh, the timeout is set to 10 minutes. The script uses run_for_benchmarks.py and generates a .csv output file.

In benchmarks.sh, we can add two boolean arguments "True" or "False" at the end of the run_for_benchmark.py calls to use dynamic reordering (3rd arg) or manual reordering (4th arg) for full bdd algorithms.

Note that create_parity_automata.sh has no timeout preset.